+++ robe-hibernate/src/main/java/io/robe/hibernate/criteria/api/cache/EntityMeta.java
@@ -8,30 +8,24 @@ import java.util.Map;
 public class EntityMeta {
     private final String identityName;
     private final Map<String, FieldMeta> fieldMap;
-    private final Map<String, String> fieldRelationMap;
-    private final Map<String, FieldMeta> relationMap;
+    private final Map<String, String> relationMap;
 
-    public EntityMeta(String identityName, Map<String, FieldMeta> fieldMap, Map<String, FieldMeta> relationMap, Map<String, String> fieldRelationMap) {
+    public EntityMeta(String identityName, Map<String, FieldMeta> fieldMap, Map<String, String> relationMap) {
         this.identityName = identityName;
         this.fieldMap = fieldMap;
         this.relationMap = relationMap;
-        this.fieldRelationMap = fieldRelationMap;
     }
 
     public String getIdentityName() {
         return identityName;
     }
 
+    public Map<String, String> getRelationMap() {
+        return relationMap;
+    }
+
     public Map<String, FieldMeta> getFieldMap() {
         return fieldMap;
-    }
-
-    public Map<String, String> getFieldRelationMap() {
-        return fieldRelationMap;
-    }
-
-    public Map<String, FieldMeta> getRelationMap(){
-        return this.relationMap;
     }
 
 }